Java Prototype of the ECG
Almost all of this prototype will have to be rewritten in Python, but I wanted to make it in Java because I'm more comfortable with it.

The main part of this code takes a three lists of points (for the P Wave, QRS Complex, and T Wave) and adds more points to curve them into looking more like a natural ECG rhythm. It also has an option to enable linear interpolation within these coordinates, making more coordinates that follow the original path to be used for the moving light on most ECG. These coordinates are also evenly spaced so this theoretical light does not speed up or slow down at certain points.

While all the code will have to be rewritten in Python, the curve algorithm, the linear interpolation (lerp) algorithm, and the finalize algorithm (which puts all the points into a path) will not have to be rewritten from scratch.